TMPFS Info Message Improvement
Modified the recently added "TMPFS free space" information attached to the "Cache Mode" option to include the current "RAM Available" (as provided by /proc/meminfo).
This allows users to see both TMPFS and RAM available space and check if the memory is becoming too low to support database operations when the Cache Mode is set to Enabled.